joystick: Concentrate info about joystick models in a single file
